replace innerVal to new implementations.
